Use unique `contentDigest` for `onCreateNode`
Currently because we use the same `contentDigest` for all nodes passing
through `onCreateNode`, they are cached by GraphQL when it returns
results. That is, if you have a query that returns three different
`Example`s, when their `contentDigest` is the same, GraphQL will simply
return three of the same record.

This commit uses the id as the digest, since id should be unique.